#26 Add de criador de token ou uuid de validacao
