+++ src/main/java/arfsoftwares/sevice/generator/token/TokenGenerator.java
@@ -1,0 +1,9 @@
+package arfsoftwares.sevice.generator.token;
+
+import arfsoftwares.data.model.Certificate;
+
+public interface TokenGenerator {
+
+	String generateToken(Certificate certificate);
+
+}
